+++ dhis2-android-app/src/main/java/org/dhis2/mobile/utils/date/filters/WeeklyPeriodFilter.java
@@ -1,14 +1,14 @@
-package org.dhis2.mobile.utils.date;
+package org.dhis2.mobile.utils.date.filters;
 
 
+import org.dhis2.mobile.utils.date.PeriodFilter;
 import org.joda.time.DateTime;
-import org.joda.time.DateTimeConstants;
 import org.joda.time.LocalDate;
 
 import java.util.Calendar;
 
-public class WeeklyWednesdayPeriodFilter extends PeriodFilter {
-    public WeeklyWednesdayPeriodFilter(DateTime startDate, DateTime endDate) {
+public class WeeklyPeriodFilter extends PeriodFilter {
+    public WeeklyPeriodFilter(DateTime startDate, DateTime endDate) {
             super(fixStartDate(startDate), fixEndDate(endDate));
         }
 
@@ -18,7 +18,7 @@ public class WeeklyWednesdayPeriodFilter extends PeriodFilter {
         }
         Calendar endDateCalendar = Calendar.getInstance();
         endDateCalendar.setTime(endDate.toDate());
-        LocalDate fixedWeek = new LocalDate(endDate.withDayOfWeek(DateTimeConstants.TUESDAY));
+        LocalDate fixedWeek = new LocalDate(endDate.withDayOfWeek(7));
         endDateCalendar.setTime(fixedWeek.toDate());
         return new DateTime(endDateCalendar.getTime());
     }
@@ -29,7 +29,7 @@ public class WeeklyWednesdayPeriodFilter extends PeriodFilter {
         }
         Calendar startDateCalendar = Calendar.getInstance();
         startDateCalendar.setTime(startDate.toDate());
-        startDateCalendar.setTime(new LocalDate(startDate.withDayOfWeek(DateTimeConstants.WEDNESDAY)).toDate());
+        startDateCalendar.setTime(new LocalDate(startDate.withDayOfWeek(1)).toDate());
         return new DateTime(startDateCalendar.getTime());
     }
 }
