+++ dhis2-android-app/src/main/java/org/dhis2/mobile/utils/date/filters/WeeklyWednesdayPeriodFilter.java
@@ -1,17 +1,17 @@
-package org.dhis2.mobile.utils.date;
+package org.dhis2.mobile.utils.date.filters;
 
 
+import org.dhis2.mobile.utils.date.PeriodFilter;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeConstants;
 import org.joda.time.LocalDate;
 
 import java.util.Calendar;
 
-public class WeeklyThursdayPeriodFilter extends PeriodFilter {
-
-    public WeeklyThursdayPeriodFilter(DateTime startDate, DateTime endDate) {
-        super(fixStartDate(startDate), fixEndDate(endDate));
-    }
+public class WeeklyWednesdayPeriodFilter extends PeriodFilter {
+    public WeeklyWednesdayPeriodFilter(DateTime startDate, DateTime endDate) {
+            super(fixStartDate(startDate), fixEndDate(endDate));
+        }
 
     private static DateTime fixEndDate(DateTime endDate) {
         if(endDate==null) {
@@ -19,7 +19,7 @@ public class WeeklyThursdayPeriodFilter extends PeriodFilter {
         }
         Calendar endDateCalendar = Calendar.getInstance();
         endDateCalendar.setTime(endDate.toDate());
-        LocalDate fixedWeek = new LocalDate(endDate.withDayOfWeek(DateTimeConstants.WEDNESDAY));
+        LocalDate fixedWeek = new LocalDate(endDate.withDayOfWeek(DateTimeConstants.TUESDAY));
         endDateCalendar.setTime(fixedWeek.toDate());
         return new DateTime(endDateCalendar.getTime());
     }
@@ -30,7 +30,7 @@ public class WeeklyThursdayPeriodFilter extends PeriodFilter {
         }
         Calendar startDateCalendar = Calendar.getInstance();
         startDateCalendar.setTime(startDate.toDate());
-        startDateCalendar.setTime(new LocalDate(startDate.withDayOfWeek(DateTimeConstants.THURSDAY)).toDate());
+        startDateCalendar.setTime(new LocalDate(startDate.withDayOfWeek(DateTimeConstants.WEDNESDAY)).toDate());
         return new DateTime(startDateCalendar.getTime());
     }
 }
